Modify and correct header

+++ header.h
@@ -1,2 +1,2 @@
-int write_header(char *input_file, char* output_file, unsigned long dict_size, unsigned long crc);
+int write_header(char *input_file, char* output_file, unsigned int dict_size, unsigned long crc);
 int read_header(char *input_file);
